+++ spine-unity/Assets/Spine/Runtime/spine-unity/Components/SkeletonMecanim.cs
@@ -126,21 +126,7 @@ namespace Spine.Unity {
 
 			skeleton.Update(deltaTime);
 
-			if (Application.isPlaying) {
-				if (applyTranslationToPhysics) {
-					Vector2 position = new Vector2(transform.position.x, transform.position.y);
-					Vector2 positionDelta = position - lastPosition;
-					positionDelta.x /= transform.lossyScale.x;
-					positionDelta.y /= transform.lossyScale.y;
-					skeleton.PhysicsTranslate(positionDelta.x, positionDelta.y);
-					lastPosition = position;
-				}
-				if (applyRotationToPhysics) {
-					float rotation = this.transform.rotation.eulerAngles.z;
-					skeleton.PhysicsRotate(0, 0, rotation - lastRotation);
-					lastRotation = rotation;
-				}
-			}
+			ApplyTransformMovementToPhysics();
 
 			ApplyAnimation();
 		}